Add GitLab API integration and project metadata
This commit introduces a GitLab token and authorization method in the `gitlab.http` file, enabling API requests to retrieve groups and projects. It also adds multiple project entries in the `projects.json` file, enhancing the metadata available for various projects, including IDs, names, paths, and visibility settings.

Additionally, significant modifications were made to the `GitLabMetricsDbContext`, including the removal of the `DimGroup` DbSet and restructuring of the `DimProject` entity. A new `SeedData` class was introduced to seed the database with project data from GitLab. The `GitLabApiService` and `IGitLabApiService` have been updated to utilize the `NGitLab` package, improving API interactions. Overall, these changes enhance the data model and streamline the database seeding process.

+++ src/Toman.Management.KPIAnalysis.ApiService/Features/GitLabMetrics/Models/Dimensions/DimUser.cs
@@ -2,7 +2,7 @@ namespace Toman.Management.KPIAnalysis.ApiService.Features.GitLabMetrics.Models.
 
 public sealed class DimUser
 {
-    public required int UserId { get; init; }
+    public required long UserId { get; init; }
     public required string Username { get; init; }
     public required string Name { get; init; }
     public required string State { get; init; }
